17_05_26_添加错题数组

+++ NZ Driving Test/NZ Driving Test/Classes/Home/Controller/YYHHomeViewController.m
@@ -87,7 +87,7 @@
 
 
     YYHButtonOnChooseTyepView *button = notification.object;
-    NSLog(@"%ld", button.tag);
+    NSLog(@"%ld", (long)button.tag);
 
     YYHCatagoriesViewController *catagory = [[YYHCatagoriesViewController alloc
                                               ] init];
